Moved spec for use language to the right place.
