+++ src/Core/Gibberish.Tests/BeAbleToDefineAThunk/ParseThunkDefinition.cs
@@ -17,14 +17,5 @@ namespace Gibberish.Tests.BeAbleToDefineAThunk
 			var result = subject.GetMatch(input, subject.TopLevelStatement);
 			Approvals.VerifyJson(result.PrettyPrint());
 		}
-
-		[Test]
-		public void just_a_uselanguage_statement_should_be_empty_parse()
-		{
-			var input = "use language fasm\r\n";
-			var subject = new ParseFasm();
-			var result = subject.GetMatch(input, subject.FasmFile);
-			Approvals.VerifyJson(result.PrettyPrint());
-		}
 	}
 }
