+++ src/Core/Gibberish.Tests/Gibberish.Tests.csproj
@@ -81,6 +81,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BeAbleToDefineAThunk\CompileThunks.cs" />
+    <Compile Include="PickTheRightLanguageForFile\ParserChooseLanguageBasedOnUseLanguageStatement.cs" />
     <Compile Include="ZzTestHelpers\ResultExtensions.cs" />
     <Compile Include="BeAbleToDefineAThunk\ParseThunkDefinition.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
