+++ src/Core/Gibberish.Tests/PickTheRightLanguageForFile/_approvals/ParserChooseLanguageBasedOnUseLanguageStatement.just_a_uselanguage_statement_should_be_empty_parse.approved.json
